+++ src/apps/concentrator/ethereum/concentrator.acrv.token-fetcher.ts
@@ -1,73 +1,50 @@
 import { Inject } from '@nestjs/common';
 
 import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
-import { Register } from '~app-toolkit/decorators';
-import { buildDollarDisplayItem } from '~app-toolkit/helpers/presentation/display-item.present';
-import { getAppImg } from '~app-toolkit/helpers/presentation/image.present';
-import { ContractType } from '~position/contract.interface';
-import { PositionFetcher } from '~position/position-fetcher.interface';
-import { AppTokenPosition } from '~position/position.interface';
-import { Network } from '~types/network.interface';
+import { PositionTemplate } from '~app-toolkit/decorators/position-template.decorator';
+import { AppTokenTemplatePositionFetcher } from '~position/template/app-token.template.position-fetcher';
+import { GetDataPropsParams, GetPricePerShareParams } from '~position/template/app-token.template.types';
 
-import { CONCENTRATOR_DEFINITION } from '../concentrator.definition';
-import { ConcentratorContractFactory } from '../contracts';
+import { AladdinCrv, ConcentratorContractFactory } from '../contracts';
 
-const appId = CONCENTRATOR_DEFINITION.id;
-const groupId = CONCENTRATOR_DEFINITION.groups.acrv.id;
-const network = Network.ETHEREUM_MAINNET;
+@PositionTemplate()
+export class EthereumConcentratorAcrvTokenFetcher extends AppTokenTemplatePositionFetcher<AladdinCrv> {
+  groupLabel = 'aCRV';
 
-export const address = '0x2b95a1dcc3d405535f9ed33c219ab38e8d7e0884'; // EIP-1967 proxy
-
-@Register.TokenPositionFetcher({ appId, groupId, network })
-export class EthereumConcentratorAcrvTokenFetcher implements PositionFetcher<AppTokenPosition> {
   constructor(
-    @Inject(APP_TOOLKIT) private readonly appToolkit: IAppToolkit,
-    @Inject(ConcentratorContractFactory) private readonly concentratorContractFactory: ConcentratorContractFactory,
-  ) {}
+    @Inject(APP_TOOLKIT) protected readonly appToolkit: IAppToolkit,
+    @Inject(ConcentratorContractFactory) protected readonly contractFactory: ConcentratorContractFactory,
+  ) {
+    super(appToolkit);
+  }
 
-  async getPositions() {
-    const multicall = this.appToolkit.getMulticall(network);
-    const baseTokens = await this.appToolkit.getBaseTokenPrices(network);
-    const underlyingToken = baseTokens.find(token => token.symbol == 'cvxCRV');
-    if (!underlyingToken) return [];
+  getContract(address: string): AladdinCrv {
+    return this.contractFactory.aladdinCrv({ address, network: this.network });
+  }
 
-    const contract = this.concentratorContractFactory.aladdinCrv({ address, network });
-    const [symbol, decimals, supplyRaw, underlyingRaw] = await Promise.all([
-      multicall.wrap(contract).symbol(),
-      multicall.wrap(contract).decimals(),
-      multicall.wrap(contract).totalSupply(),
-      multicall.wrap(contract).totalUnderlying(),
-    ]);
+  async getAddresses() {
+    return ['0x2b95a1dcc3d405535f9ed33c219ab38e8d7e0884'];
+  }
 
-    const supply = Number(supplyRaw) / 10 ** decimals;
-    const pricePerShare = Number(underlyingRaw) / Number(supplyRaw);
-    const price = underlyingToken.price * pricePerShare;
-    const liquidity = price * supply;
+  async getUnderlyingTokenAddresses() {
+    return ['0x62b9c7356a2dc64a1969e19c23e4f579f9810aa7'];
+  }
 
-    // TODO: add additional properties from https://concentrator-api.aladdin.club/apy/
-    const label = `Concentrator aCRV`;
-    const images = [getAppImg(appId)];
+  async getPricePerShare({ appToken, contract }: GetPricePerShareParams<AladdinCrv>) {
+    const reserveRaw = await contract.totalUnderlying();
+    const reserve = Number(reserveRaw) / 10 ** appToken.tokens[0].decimals;
+    return reserve / appToken.supply;
+  }
 
-    const token: AppTokenPosition = {
-      type: ContractType.APP_TOKEN,
-      appId,
-      groupId,
-      address,
-      network,
-      symbol,
-      decimals,
-      supply,
-      tokens: [underlyingToken],
-      price,
-      pricePerShare,
-      dataProps: { liquidity },
-      displayProps: {
-        label,
-        images,
-        statsItems: [{ label: 'Liquidity', value: buildDollarDisplayItem(liquidity) }],
-      },
-    };
+  async getLiquidity({ appToken }: GetDataPropsParams<AladdinCrv>) {
+    return appToken.supply * appToken.price;
+  }
 
-    return [token];
+  async getReserves({ appToken }: GetDataPropsParams<AladdinCrv>) {
+    return [appToken.pricePerShare[0] * appToken.supply];
+  }
+
+  async getApy(_params: GetDataPropsParams<AladdinCrv>) {
+    return 0;
   }
 }