+++ src/apps/synthetix/contracts/index.ts
@@ -5,10 +5,12 @@ import { ContractFactory } from '~contract/contracts';
 import { Network } from '~types/network.interface';
 
 import { SynthetixAddressResolver__factory } from './ethers';
+import { SynthetixExchangeRates__factory } from './ethers';
 import { SynthetixLoan__factory } from './ethers';
 import { SynthetixNetworkToken__factory } from './ethers';
 import { SynthetixRewards__factory } from './ethers';
 import { SynthetixSummaryUtil__factory } from './ethers';
+import { SynthetixSynthToken__factory } from './ethers';
 
 // eslint-disable-next-line
 type ContractOpts = { address: string; network: Network };
@@ -22,6 +24,9 @@ export class SynthetixContractFactory extends ContractFactory {
   synthetixAddressResolver({ address, network }: ContractOpts) {
     return SynthetixAddressResolver__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
+  synthetixExchangeRates({ address, network }: ContractOpts) {
+    return SynthetixExchangeRates__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  }
   synthetixLoan({ address, network }: ContractOpts) {
     return SynthetixLoan__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
@@ -34,10 +39,15 @@ export class SynthetixContractFactory extends ContractFactory {
   synthetixSummaryUtil({ address, network }: ContractOpts) {
     return SynthetixSummaryUtil__factory.connect(address, this.appToolkit.getNetworkProvider(network));
   }
+  synthetixSynthToken({ address, network }: ContractOpts) {
+    return SynthetixSynthToken__factory.connect(address, this.appToolkit.getNetworkProvider(network));
+  }
 }
 
 export type { SynthetixAddressResolver } from './ethers';
+export type { SynthetixExchangeRates } from './ethers';
 export type { SynthetixLoan } from './ethers';
 export type { SynthetixNetworkToken } from './ethers';
 export type { SynthetixRewards } from './ethers';
 export type { SynthetixSummaryUtil } from './ethers';
+export type { SynthetixSynthToken } from './ethers';